📝 docs(README.md): update socialMedia attribute description to use array of objects instead of JSON string
The `socialMedia` attribute in the digital business card component now expects an array of objects instead of a JSON string. Each object should have a `platform` and a `url` property. This change allows for more flexibility and easier customization of social media links.

Example usage:

```json
[
    { "platform": "facebook", "url": "https://facebook.com/johndoe" },
    { "platform": "twitter", "url": "https://twitter.com/johndoe" },
    { "platform": "mastodon", "url": "https://mastodon.social/johndoe" }
]
```

This change improves the clarity and maintainability of the code.